# Improve glyph copy performance by using read offset