Forgot the offset parameter exists.

```
|                             Method |     Mean |     Error |    StdDev |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|----------------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|----------:|
|                   TimedExpiryCache | 19.38 ms | 0.0493 ms | 0.0437 ms |  93.7500 |  62.5000 |  62.5000 |   4.14 MB |
|                   RawSaveNoCaching | 26.17 ms | 0.2551 ms | 0.2262 ms | 750.0000 | 625.0000 | 562.5000 |    4.9 MB |
|                 RawSaveNoCachingV2 | 22.55 ms | 0.0501 ms | 0.0468 ms | 750.0000 | 625.0000 | 562.5000 |    4.9 MB |
|            TimedExpiryCacheTrimmed | 16.37 ms | 0.0679 ms | 0.0635 ms | 187.5000 | 187.5000 | 156.2500 |   3.29 MB |
|            RawSaveNoCachingTrimmed | 24.13 ms | 0.1368 ms | 0.1213 ms | 968.7500 | 968.7500 | 968.7500 |   4.05 MB |
|          RawSaveNoCachingTrimmedV2 | 20.55 ms | 0.0810 ms | 0.0633 ms | 968.7500 | 968.7500 | 968.7500 |   4.05 MB |
```

`readBuffer` late initialisation is not an optimisation but a correctness change (ie. if the user was using atlases larger than 1024px with super huge glyphs, it may have overflown).